ia_power_nr/requirements.txt
pre-commit==2.8.2
numpy==1.19.4
pandas==1.1.4
scipy==1.5.4
matplotlib==3.3.3
seaborn==0.11.0
statsmodels==0.12.1
scikit-learn==0.23.2
pyyaml==5.3.1
black==20.8b1
flake8==3.8.4
ipython==7.19.0
jupyter==1.0.0
jupyterlab==2.2.9
python-dotenv==0.15.0
pymc3==3.11.1
pystan==2.19.1.1
google-cloud-bigquery==1.24.0
google-cloud-bigquery-storage==2.0.0
pyarrow==2.0.0
pytest==6.2.2
python-dotenv==0.15.0
pymc3==3.11.1
pystan==2.19.1.1

